Don't create a util package - move the functionality to 'internal'
